+++ src/detour/x86_64/x86_64.cc
@@ -1,5 +1,5 @@
 #include "x86_64.h"
-#include "detour_impl.h"
+#include "detour/detour_impl.h"
 
 #include <spud/detour.h>
 #include <spud/memory/protection.h>
